backend/requirements.txt
asgiref==3.5.2
Django==3.2.20
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
sqlparse==0.4.4
tzdata==2023.3